app/api/requirements.txt
azure-functions==1.21.3
azure-storage-blob==12.24.0
azure-identity==1.19.0
python-jose[cryptography]==3.3.0
httpx==0.28.0
